requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
sqlalchemy==2.0.30
pydantic==2.7.1
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.9
bcrypt==4.0.1
